Big overhaul starting
It has been requested to make the VI fool and/or futureproof. The same sensors will not be wired to the same (physical) terminals every time, and sometimes not every terminal is in use while other times they will be. VI will be overhauled with this in mind - read out every single terminal every run, and specify which sensor is which, where needed.