v2, more robust
Added case structure just before the output, that checks if the array passed has 8 entries. If it doesnt, something probably went wrong, and the output might be all zeroes, losing all information. 
Now, if the output array is not size 8, an array of eight 1's is passed to not alter VMon outputs.