More flexibility and PT1000 integration
Changed some more constants to values like the sum of some array sizes, to keep plots from becoming mangled when a different amount of RTDs are connected. Still need to rework plotting a bit more, but don't intend to make dummy data for the PT1000s as the reading module hasn't arrived yet, so that will have to wait.