Includes passband filter for det temp
As title. On "Plot" frame, the average was taken from all NI 9226 data readouts. If a terminal is unconnected but being read out, data is shown as either 1300 deg C or -240 deg C. The passband filter deletes these entries from the array and passes the correct array on, to average only the "normal" data points.